social_auth_django/pyproject.toml
[tool.poetry]
name = "social-auth-django"
version = "0.1.0"
description = ""
authors = ["ngoclinh8123 <ngoclinh8123@gmail.com>"]
readme = "README.md"
packages = [{include = "social_auth_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-social-oauth2 = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
